lab6/bot/requirements.txt
charset-normalizer
pyTelegramBotAPI
requests
tortoise_orm
tortoise_orm[asyncpg]
urllib3
pydantic
secret
python-decouple
aiohttp
MoviePy
redis
redis[hiredis]
asyncio-redis
schedule
apscheduler
numpy
cairosvg
svgwrite
yookassa
var_dump
Pillow==9.5.0